mb/google/skolas: Skip locking for GPP_F14 GPIO
There is an existing issue for skolas boards where board wakes up
from shutdown immediately due to touchpad wake signal.

This issue was root caused to the patch which was setting GPE_EN
bits for the GPIOs before locking.
Ref: commit 38b8bf02d820 ("intelblocks: Add function to program GPE_EN before GPIO locking")

Later issue was found to be with GPP_F14 configuration for skolas
boards. While shutting down, kernel is not able to disable IRQ for
touchpad due to GPE_EN register getting locked and it is preventing
shutdown of the board.

This patch skips the locking for GPP_F14 to allow kernel to
configure it later.

BUG=b:254064671
BRANCH=None
TEST=Shutdown works on Skolas board with the patch.

Nissa Bug: 234097956

Signed-off-by: Maulik Vaghela <maulikvaghela@google.com>
Change-Id: I09cf1af1f5ab11b06073755374ee8a306984d557
Reviewed-on: https://review.coreboot.org/c/coreboot/+/72426
Reviewed-by: Subrata Banik <subratabanik@google.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>