Documentation: Update acronyms list
This change adds some new acronyms to the list, clarifies a couple of
points, and fixes a couple formatting issues.

I was planning on leaving this open for a bit and continuing to add
to the patch.  If anyone else wants to help, please feel free to update
this patch as you see fit.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Change-Id: I07212849640e8ef14e3c4a41ade29498a4578bc6
Reviewed-on: https://review.coreboot.org/c/coreboot/+/72923
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <felixsinger@posteo.net>